rename logger files
